refactor(completion): перенести CompletionData в пакет completion по образцу CodeLensData
